use build tags to allow use of unsafe
